pyproject.toml
[tool.poetry]
name = "fast-cached-api"
version = "0.1.0"
description = ""
authors = ["Anderson Rodrigues <eusouanderson@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
motor = "^3.6.0"
redis = {extras = ["asyncio"], version = "^5.2.1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
packaging = "^24.2"
pydantic-settings = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
